Hide actions on a deleted site
